AnimeAvatarGenerator/requirements.txt
torch >= 1.0.0
torchvision
argparse
pillow